refactor: domain 패키지, view 패키지, controller 패키지 분리 - domain 패키지 객체가 view 패키지 객체에 의존하지 않도록 분리
