+++ src/main/java/RacingGame.java
@@ -1,4 +1,4 @@
-import racingcar.Game;
+import controller.Game;
 
 public class RacingGame {
 	public static void main(String[] args) {
