+++ src/main/java/racingcar/Cars.java
@@ -1,7 +1,5 @@
 package racingcar;
 
-import io.PrintView;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -16,11 +14,18 @@ public class Cars {
         }
     }
 
-    public void move() {
-        cars.forEach(car -> {
-            car.move();
-            PrintView.printPosition(car);
-        });
+    public List<Car> move() {
+        cars.forEach(Car::move);
+
+        return cars;
+    }
+
+    public List<Car> getWinners() {
+        int maxPosition = getMaxPosition();
+
+        return cars.stream()
+                .filter(car -> car.isEqualPosition(maxPosition))
+                .collect(Collectors.toList());
     }
 
     private int getMaxPosition() {
@@ -29,14 +34,4 @@ public class Cars {
                 .max()
                 .orElse(0);
     }
-
-    public String getWinnerName() {
-        int maxPosition = getMaxPosition();
-
-        return cars.stream()
-                .filter(car -> car.isEqualPosition(maxPosition))
-                .map(Car::getCarName)
-                .collect(Collectors.joining(","));
-    }
-
 }
