+++ src/main/java/view/InputView.java
@@ -1,4 +1,4 @@
-package io;
+package view;
 
 import java.util.Scanner;
